Fix issue with time_t being a uint64_t on some platforms

+++ app_wsbrd/ipv6/ipv6_neigh_storage.c
@@ -69,7 +69,7 @@ void ipv6_neigh_storage_save(struct ipv6_neighbour_cache *cache, const uint8_t *
         fprintf(nvm->file, "ipv6[%d] = %s\n", i, ipv6_str);
         fprintf(nvm->file, "lifetime[%d] = %u\n", i, cur->lifetime_s);
         fprintf(nvm->file, "# %s\n", time_str);
-        fprintf(nvm->file, "expiration[%d] = %lu\n", i, ts);
+        fprintf(nvm->file, "expiration[%d] = %ju\n", i, (uintmax_t) ts);
         i++;
     }
